+++ Frontend/Main_GUI.java
@@ -1,3 +1,6 @@
+import Main_Game.Main_Game_Page;
+import Main_Menu.Main_Menu_Page;
+
 public class Main_GUI {
     public static void main(String[] args) {
         new Main_Game_Page();
