+++ Frontend/Main_Menu/Main_Menu_Page.java
@@ -1,3 +1,5 @@
+package Main_Menu;
+
 public class Main_Menu_Page {
 
 }
